+++ docs/CHANGELOG.rst
@@ -30,6 +30,7 @@ Changed
   processes is no longer required
 - Freeze docutils package version to 0.15.2 because Sphinx has
   problems parsing development version numbers
+- Support Slamdunk/Alleyoop processes in MultiQC
 
 ===================
 24.0.0 - 2019-11-15